Update movie.html
Changed the filtering for what is displayed in searches. Now, anything with "short" in the title or genre is filtered out, as well as anything with "N/A" as a rating or director, non-case sensitive. Also added a cap of 30 movies per search to avoid hitting the rate limit as quickly as before.