Update watchparty_result.html
Changed it to extend layout.html instead of base.html and improved how it searches for random movies. Now, it searches using TMDb because it is better equipped to handle searches without a title and then uses OMDb to display the movie poster. May consider switching it to use TMDb for posters as well at a later date.